requirements.txt
asciitree==0.3.3
click==7.1.2
dask==2.18.0
fastapi[all]==0.55.1
fasteners==0.15
mercantile==1.1.4
monotonic==1.5
numcodecs==0.6.3
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
pypng==0.0.20
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
toolz==0.10.0
xarray==0.15.1
zarr==2.3.2
